+++ src/main/kotlin/de/mariushoefler/flutterenhancementsuite/exceptions/PubExceptions.kt
@@ -11,3 +11,27 @@ class GetCurrentPackageVersionException(p: String) :
 class PubApiCouldNotBeReachedException(p: Exception) : Exception("Pub api could not be reached: $p")
 
 class PubApiUnknownFormatException(p: JsonSyntaxException) : Exception("Unexpected response from Pub api: $p")
+
+class AutoDisposeChangeNotifierProviderFamilyThisNameIsWayTooLongToBeDisplayed<T, R>()
+
+fun foo (a: AutoDisposeChangeNotifierProviderFamilyThisNameIsWayTooLongToBeDisplayed<String, String>, b: String) {
+
+}
+
+fun boo() {
+    foo(w, b)
+}
+
+class Fee {
+    private val pubApiService by lazy {
+        Tee.create()
+    }
+}
+
+class Tee {
+    companion object {
+        fun create(): AutoDisposeChangeNotifierProviderFamilyThisNameIsWayTooLongToBeDisplayed<String, String> {
+            return AutoDisposeChangeNotifierProviderFamilyThisNameIsWayTooLongToBeDisplayed()
+        }
+    }
+}